requirements.txt
python-telegram-bot==21.6
yt-dlp==2026.3.17
